+++ packages/flood-runner/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-canary.15](https://github.com/flood-io/element/compare/v2.0.0-canary.14...v2.0.0-canary.15) (2021-01-06)
+
+
+### Features
+
+* show screenshot on terminal (iTerm) ([8fa8c88](https://github.com/flood-io/element/commit/8fa8c88e6f86477a1adc95b2de4ed710d70d7531))
+
+
+
+
+
+# [2.0.0-canary.14](https://github.com/flood-io/element/compare/v2.0.0-canary.13...v2.0.0-canary.14) (2021-01-04)
+
+**Note:** Version bump only for package @flood/element-flood-runner
+
+
+
+
+
 # [2.0.0-canary.13](https://github.com/flood-io/element/compare/v2.0.0-canary.12...v2.0.0-canary.13) (2020-12-28)
 
 **Note:** Version bump only for package @flood/element-flood-runner